feat: initialize flake
This was pulled from https://github.com/justinrubek/lockpad. The premise
is simple: this contains some scripts that let you start up the pinned
version of postgres when using the project. This is being placed here as
an archive of the concept as it will not be used any longer.